Set creation is no longer all done at the end
Now every 20 uploads the sets are updated. This means you can see some
progress on the webpage. No modification of sets that have not been
altered in this upload session is done. This means the time taken to
upload photos is reduced, as the script no longer re-sets all the
previously uploaded photos.