Add a check for the configuration file
This raises an IOError if the config file is not found in one of the
locations it is expected to be. e.g. running from a $HOME directory
the error would be:

Missing configuration file: /home/user/.uploadr.ini /home/user/uploadr.ini

At least this way you know where the scripts are looking without having
to code dive. Fixes #8.