Add check for null sets.
Prevent tags2set from attempting to create an empty set. Previously, it would
attempt to create a set based on the image's directory. If that directory were
the top-level directory (as specified in the ini), string sanitization would
remove the / and attempt to create an empty set ''. This would through an error.

Now, if an empty set is detected, _creatSet simply returns.